requirements.txt
deta==1.1.0a2
matplotlib==3.5.2
matplotlib-inline==0.1.3
numpy==1.23.1
numpy-financial==1.0.0
pandas==1.5.0
plotly==5.8.0
plotly-express==0.4.1
PuLP==2.6.0
scipy==1.9.0
seaborn==0.11.2
sklearn==0.0
spacy==3.4.1
spacy-streamlit==1.0.4
SQLAlchemy==1.4.39
st-card-component==0.0.10
statsmodels==0.13.2
streamlit==1.13.0
streamlit-aggrid==0.2.3.post2
streamlit-card==0.0.3
streamlit-folium==0.6.14
streamlit-on-Hover-tabs==1.0.1
toml==0.10.2
tqdm==4.64.1
transformers
openpyxl
mysql.connector
gspread
pip==22.2.2
datetime
psycopg2-binary
sqlalchemy
gspread
pathlib
PuLP==2.6.0
sklearn==0.0
nltk
goose3
networkx
openai
